moving js comment functionality to plugin
